Changes include
(1) TILDA___Docs.PATIENTS.html -> _tilda.Patients.html

(2) _tilda.Patients.graph.json should include both the rendered SVG for the Gen Utility, AND the positions saved.

(3) Clean up the views to only show the views and their direct dependent tables.

(4) get back to using GraphViz for initial layout for default rendering.

(5) In the docs, disable the hover styling

(6) In the docs AND in the tool, use double-click to toggle on/off the "related object highlighting" features, and in the tool, remove the Reset button.

(7) break up the SVG line into multiple lines. so we can view the source in a browser. Right now, a line of 373,000 characters is created which crashed the browser when viewing in source mode.

(8) adding "Xyz Schema" at the top of the generated final HTML document, along with the description (I will add a "description" property at the top level JSON)