#144 documentation updates
Cleanup of migration to do column null checks after type checks. It's
possible for a column to change from let's say INTEGER to STRING, and
also have a new default value and be not-null... The order of the
migrate scripts could cause an error in this case as it would try to set
up the default value first (which would fail), and then do a type
change.