#237
More work to add batch process to table updated. I have the process more
or less working for the AlterTableColumnType, but there are issues
unrelated to the changes I've made that I'll need to review.

Shelving this work for higher priority work. I believe the pattern is a
good starting point for what needs to happen.