#444 java runtime data and factory API cleanup
Major cleanup of all APIs to have methods that start with a lowercase.
This breaks backward compatibility but believe it's for the good of the
project.

Most utility classes, interfaces, and generated DATA/FACTORY/JSON
classes affected, for things as basic as the LookupXXX(), Read() or
Write() methods (to lookupXXX, read, write...).