#567 savepoint exception when starting application with validation:true
There appears to have been 2 problems:
- The main savepoint management logic in Connection called the API "add"
instead of "push", which caused the behavior to not follow the Stack
semantics that we need.
- When skipValidation is set to false, the full validation occurs and we
perform a full migration test. The main logic in
DDLDependencyPreManagement.isNeeded creates a savepoint without in some
cases releasing it. We updated the code to do proper releasing in all
code paths.