#767 While outputting migration SQL, some SQL still executes
We created the Rehearsal mode to output the migration workflow to an SQL
file. However, as part of migration, we do execute some SQL to check
things, for instance, whether a table has any rows before adding a
non-null column, or checking the size of a STRING column before
shrinking in.

We need to capture those in rehearsal mode so that they don't execute:
they will execute as par of the proper migration itself.

Finally, took advantage to do some optimizations. For example, we do
count(*) on a lot of these checks, when we should only check for the
existence of at least one row. So queries such as:
```sql
select count(*) from TableX
```
Can be changed to
```sql
select 1 from TableX limit 1
```
This approach significantly improves over all performance of the
migration.

Finally, also added getRepeatableField* helpers to BQHelper.